Update server and client folder
