+++ client/src/App.jsx
@@ -2,18 +2,24 @@ import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { CardThunk } from "./redux/CardSlice";
 import Navbar from "./Components/Navbar";
+import Product from "./Components/Product";
+import CarouselCard from "./Components/CarouselCard";
 
 function App() {
-  const store = useSelector((state) => state.card.value);
+  const store = useSelector((state) => state.card.data);
   const dispatch = useDispatch();
 
   useEffect(() => {
-    if (store === 0) dispatch(CardThunk());
-  }, []);
+    if (store.length === 0) {
+      dispatch(CardThunk());
+    }
+  });
 
   return (
     <div className="App">
-      <Navbar />
+      {/* <Navbar /> */}
+      <CarouselCard/>
+      <Product/>
     </div>
   );
 }
