+++ client/src/styles/Navbar.css
@@ -1,0 +1,32 @@
+.sidebar {
+  height: 100vh;
+  width: calc(300px + 0vw);
+  background-color: #2f4050;
+}
+
+.sidebarList .row {
+  font-size: 1.5rem;
+  padding: 2rem 0;
+  border: 1px solid rgba(0, 0, 0, 0.3803921569);
+  list-style-type: none;
+  display: flex;
+  color: white;
+  align-items: center;
+  justify-content: center;
+}
+.sidebarList .row:hover {
+  cursor: pointer;
+  background-color: #293846;
+}
+.sidebarList #active {
+  background-color: #1b2b3a;
+}
+
+.row #icon {
+  flex: 30%;
+  display: grid;
+  place-items: center;
+}
+.row #title {
+  flex: 70%;
+}/*# sourceMappingURL=Navbar.css.map */
