+++ client/src/styles/Product.css
@@ -1,0 +1,4 @@
+.card-container {
+  display: grid;
+  grid: 100%/auto auto auto;
+}/*# sourceMappingURL=Product.css.map */
